refactor(webhooks): import getCarbonServiceRole directly
The lazy `await import("@carbon/auth/client.server")` was an outlier -- every
sibling event handler (audit, sync, search, embedding) imports
getCarbonServiceRole at module scope. It existed only so the contract tests
could import this file, since client.server validates the full server env on
load.

Solved the way the codebase already solves it: pure logic moves to its own
peer module. audit.ts imports getCarbonServiceRole directly and gets its
tested logic from diff.ts; webhook.ts now does the same with webhook-body.ts.
The contract tests move with the function (webhook-body.test.ts), unchanged
and still passing.
